All ContextAPI state and actions has deleted
