+++ client/src/App.jsx
@@ -1,16 +1,15 @@
+import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import { lazy, Suspense } from "react";
 import { BrowserRouter, Navigate, Route, Routes } from "react-router-dom";
+import { ToastContainer } from "react-toastify";
+import "react-toastify/dist/ReactToastify.css";
 import City from "./components/City";
 import CityList from "./components/CityList";
 import CountryList from "./components/CountryList";
 import Form from "./components/Form";
 import Map from "./components/Map";
 import SpinnerFullPage from "./components/SpinnerFullPage";
-import { CitiesProvider } from "./contexts/CityContext";
 import ProtectedRoute from "./pages/ProtectedRoute";
-import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
-import { toast, ToastContainer } from "react-toastify";
-import "react-toastify/dist/ReactToastify.css";
 
 const Homepage = lazy(() => import("./pages/Homepage"));
 const Product = lazy(() => import("./pages/Product"));
@@ -31,40 +30,37 @@ const queryClient = new QueryClient({
 function App() {
   return (
     <QueryClientProvider client={queryClient}>
-      <CitiesProvider>
-        <BrowserRouter>
-          <Suspense fallback={<SpinnerFullPage />}>
-            <Routes>
-              <Route index element={<Homepage />} />
-              <Route path="product" element={<Product />} />
-              <Route path="pricing" element={<Pricing />} />
-              <Route path="login" element={<Login />} />
-              <Route path="register" element={<Register />} />
+      <BrowserRouter>
+        <Suspense fallback={<SpinnerFullPage />}>
+          <Routes>
+            <Route index element={<Homepage />} />
+            <Route path="product" element={<Product />} />
+            <Route path="pricing" element={<Pricing />} />
+            <Route path="login" element={<Login />} />
+            <Route path="register" element={<Register />} />
 
-              <Route
-                path="app"
-                element={
-                  <ProtectedRoute>
-                    <AppLayout />
-                  </ProtectedRoute>
-                  // <AppLayout />
-                }
-              >
-                {/* <Route index element={<Navigate to="cities" replace />} /> */}
-                <Route index element={<Navigate to="cities" replace />} />
-                <Route path="cities" element={<CityList />} />
+            <Route
+              path="app"
+              element={
+                <ProtectedRoute>
+                  <AppLayout />
+                </ProtectedRoute>
+              }
+            >
+              {/* <Route index element={<Navigate to="cities" replace />} /> */}
+              <Route index element={<Navigate to="cities" replace />} />
+              <Route path="cities" element={<CityList />} />
 
-                <Route path="cities/:id" element={<City />} />
-                <Route path="countries" element={<CountryList />} />
-                <Route path="form" element={<Form />} />
-                <Route path="map" element={<Map />} />
-              </Route>
-              <Route path="*" element={<PageIsNotFound />} />
-            </Routes>
-          </Suspense>
-          <ToastContainer />
-        </BrowserRouter>
-      </CitiesProvider>
+              <Route path="cities/:id" element={<City />} />
+              <Route path="countries" element={<CountryList />} />
+              <Route path="form" element={<Form />} />
+              <Route path="map" element={<Map />} />
+            </Route>
+            <Route path="*" element={<PageIsNotFound />} />
+          </Routes>
+        </Suspense>
+        <ToastContainer />
+      </BrowserRouter>
     </QueryClientProvider>
   );
 }
