+++ client/src/components/Sidebar.jsx
@@ -1,22 +1,38 @@
+import { useEffect } from "react";
 import { FaGripLines } from "react-icons/fa";
 import { FaGripLinesVertical } from "react-icons/fa6";
+import { useDispatch, useSelector } from "react-redux";
 import { Outlet } from "react-router-dom";
-import { useCities } from "../contexts/CityContext";
+import { setIsSidebarOpen } from "../redux/sidebarSlice";
 import AppNav from "./AppNav";
 import Logo from "./Logo";
 import styles from "./Sidebar.module.css";
 
 function Sidebar() {
-  const { showMap, setShowMap, resizeableRef } = useCities();
+  const { isSideBarOpen } = useSelector((state) => state.sidebar);
+  const dispatch = useDispatch();
+
+  useEffect(() => {
+    const handleResize = () => {
+      if (window.innerWidth > 1024) {
+        dispatch(setIsSidebarOpen(true));
+      }
+    };
+    handleResize();
+    window.addEventListener("resize", handleResize);
+
+    return () => {
+      window.removeEventListener("resize", handleResize);
+    };
+  }, [dispatch]);
 
   return (
     <div
       className={styles.sidebar}
-      ref={resizeableRef}
-      style={{ height: showMap ? "100dvh" : "8rem" }}
+      style={{ height: isSideBarOpen ? "100dvh" : "8rem" }}
     >
       <Logo className={"position"} />
-      {showMap && (
+      {isSideBarOpen && (
         <>
           <AppNav />
           <Outlet />
@@ -30,7 +46,7 @@ function Sidebar() {
       <div className={styles.devider}>
         <FaGripLines
           className={styles.doubleLine}
-          onClick={() => setShowMap((isOpen) => !isOpen)}
+          onClick={() => dispatch(setIsSidebarOpen(!isSideBarOpen))}
         />
       </div>
       <div className={styles.deviderVertical}>
