+++ client/src/components/Sidebar.module.css
@@ -3,7 +3,7 @@
   background-color: var(--color-dark--1);
   padding: 3rem 5rem 3.5rem 5rem;
   flex-basis: 56rem;
-
+  z-index: 1001;
   display: flex;
   flex-direction: column;
   align-items: center;
