--- client/src/contexts/CityContext.jsx
@@ -1,198 +1,0 @@
-import axios from "axios";
-import {
-  createContext,
-  useCallback,
-  useContext,
-  useEffect,
-  useReducer,
-  useRef,
-  useState,
-} from "react";
-
-const BASE_URL = "http://localhost:9000";
-
-const CitiesContext = createContext();
-
-const initialState = {
-  cities: [],
-  isLoading: false,
-  currentCity: {},
-  error: "",
-  isLaptop: false,
-};
-
-const reducer = (state, action) => {
-  switch (action.type) {
-    case "loading":
-      return {
-        ...state,
-        isLoading: true,
-      };
-    case "cities/loaded":
-      return {
-        ...state,
-        isLoading: false,
-        cities: action.payload,
-      };
-
-    case "city/loaded":
-      return {
-        ...state,
-        isLoading: false,
-        currentCity: action.payload,
-      };
-    case "city/created":
-      return {
-        ...state,
-        isLoading: false,
-        cities: [...state.cities, action.payload],
-        currentCity: action.payload,
-      };
-    case "city/deleted":
-      return {
-        ...state,
-        isLoading: false,
-        cities: state.cities.filter((city) => city.id !== action.payload),
-        currentCity: {},
-      };
-    case "rejected":
-      return {
-        ...state,
-        isLoading: false,
-        error: action.payload,
-      };
-
-    default:
-      throw new Error("Unknown action type!");
-  }
-};
-
-function CitiesProvider({ children }) {
-  const [{ cities, isLoading, currentCity, error }, dispatch] = useReducer(
-    reducer,
-    initialState
-  );
-  const [showMap, setShowMap] = useState(true);
-
-  const resizeableRef = useRef(null);
-
-  useEffect(() => {
-    const handleResize = () => {
-      if (window.innerWidth > 1024) {
-        setShowMap(true);
-      }
-    };
-    handleResize();
-    window.addEventListener("resize", handleResize);
-
-    return () => {
-      window.removeEventListener("resize", handleResize);
-    };
-  }, [setShowMap]);
-
-  useEffect(() => {
-    const fetchCities = async () => {
-      dispatch({ type: "loading" });
-      try {
-        const res = await axios.post(
-          "/api/users/get-cities",
-          {},
-          {
-            headers: {
-              Authorization: `Bearer ${localStorage.getItem("token")}`,
-            },
-          }
-        );
-        const data = await res.data.cities;
-        dispatch({ type: "cities/loaded", payload: data });
-      } catch {
-        dispatch({
-          type: "rejected",
-          payload: "There was an error loading cities...",
-        });
-      }
-    };
-    fetchCities();
-  }, []);
-
-  const getCity = useCallback(
-    async function getCity(id) {
-      if (Number(id) === currentCity.id) return;
-      dispatch({ type: "loading" });
-      try {
-        const res = await fetch(`${BASE_URL}/cities/${id}`);
-        const data = await res.json();
-        dispatch({ type: "city/loaded", payload: data });
-      } catch {
-        dispatch({
-          type: "rejected",
-          payload: "There was an error geting city...",
-        });
-      }
-    },
-    [currentCity.id]
-  );
-
-  async function createCity(newCity) {
-    dispatch({ type: "loading" });
-    try {
-      const res = await fetch(`${BASE_URL}/cities`, {
-        method: "POST",
-        body: JSON.stringify(newCity),
-        headers: {
-          "Content-Type": "application/json",
-        },
-      });
-      const data = await res.json();
-      dispatch({ type: "city/created", payload: data });
-    } catch {
-      dispatch({
-        type: "rejected",
-        payload: "There was an error creating city...",
-      });
-    }
-  }
-
-  async function deleteCity(id) {
-    dispatch({ type: "loading" });
-    try {
-      const res = await fetch(`${BASE_URL}/cities/${id}`, {
-        method: "DELETE",
-      });
-      dispatch({ type: "city/deleted", payload: id });
-    } catch {
-      dispatch({
-        type: "rejected",
-        payload: "There was an error deleting city...",
-      });
-    }
-  }
-
-  return (
-    <CitiesContext.Provider
-      value={{
-        cities,
-        isLoading,
-        currentCity,
-        createCity,
-        getCity,
-        deleteCity,
-        error,
-        showMap,
-        setShowMap,
-        resizeableRef,
-      }}
-    >
-      {children}
-    </CitiesContext.Provider>
-  );
-}
-
-const useCities = () => {
-  const context = useContext(CitiesContext);
-  if (context === undefined)
-    throw new Error("CitiesContext values was used out of CitiesProvider!");
-  return context;
-};
-
-export { CitiesProvider, useCities };
