+++ client/src/redux/store.js
@@ -1,10 +1,12 @@
 import { combineReducers, configureStore } from "@reduxjs/toolkit";
 import userSlice from "./userSlice";
 import currentCitySlice from "./currentCitySlice";
+import sidebarSlice from "./sidebarSlice";
 
 const rootReducer = combineReducers({
   user: userSlice,
   currentCity: currentCitySlice,
+  sidebar: sidebarSlice,
 });
 
 const store = configureStore({
